requirements.txt
git+https://github.com/openai/swarm.git
python-dotenv>=1.0.1
yfinance>=0.2.44